+++ packages/ui/src/passthroughs/SplitButton.pt.ts
@@ -1,4 +1,5 @@
 import { SplitButtonPassThroughMethodOptions } from "primevue/splitbutton";
+import { TieredMenuPassThroughMethodOptions } from "primevue";
 
 export default {
   splitbutton: {
@@ -31,9 +32,24 @@ export default {
       })
     },
     pcMenu: {
-      rootList: 'dropdown-menu show w-100',
-      item: 'dropdown-item',
-      separator: 'border-bottom w-100'
+      rootList: ({instance}: TieredMenuPassThroughMethodOptions<any>) => {
+        const classes = ['dropdown-menu show']
+        // We do some calculations here to figure out if the dropdown expands beyond the page width
+        // this can happen often if the splitbutton is positioned towards the right of the page and then gets fed in long worded options
+        // without this tweak, the page will expand outwards, which looks wrong
+        // this fix adds another css class which aligns the dropdown to the right side, causing long options to expand into the page, instead of outwards
+        // in every other scenario, the dropdown expands as it did before
+        const ul = document.querySelector(`ul[teleported-from=${instance.$el.parentElement.id}]`)
+        if(ul?.parentElement?.offsetLeft + ul?.offsetWidth > window.innerWidth)
+          classes.push('dropdown-menu-end')
+        return {
+          'teleported-from': instance.$el.parentElement.id,
+          'data-bs-popper': '',
+          class: classes
+        }
+      },
+      itemLink: 'dropdown-item',
+      separator: 'border-bottom'
     }
   }
 }